Seperate testlib module into directory
