+++ noline/src/testlib/csi.rs
@@ -1,0 +1,7 @@
+pub const UP: &str = "\x1b[A";
+pub const DOWN: &str = "\x1b[B";
+pub const LEFT: &str = "\x1b[D";
+pub const RIGHT: &str = "\x1b[C";
+pub const HOME: &str = "\x1b[1~";
+pub const DELETE: &str = "\x1b[3~";
+pub const END: &str = "\x1b[4~";
